found another cue file
